requirements.txt
flask==3.1.0
python-dotenv==1.1.0
google-generativeai==0.8.5
yt-dlp==2025.1.26
